requirements.txt
streamlit
PyMuPDF
tiktoken
langchain
sentence-transformers
faiss-cpu
transformers
python-dotenv
PyPDF2
openpyxl
